Added a new feature that eliminates the problem with the USB serial drivers conflict. Previously, if the kernel modules ftdi_sio and usbserial were loaded, psi46expert could not open a connection to the testboard. One had to unload or permanently blacklist the two modules in order to work with the software (with libftdi-1.0 that is). This caused problems with other devices which required the usbserial driver. The change in this commit tries to detach the kernel driver from the selected testboard when the initial attempt to open the device failed. When the detaching succeeds another attempt at opening the device is made.
This change works only for Linux installations. The kernel driver
detach function will give an error on other platforms, but this
happens only after the first attempt to open the device already
occured. Therefore the impact of this additional code is minimal.

A bugfix is made along with the added feature: An errorneous
statement was corrected which caused some tests of the FullTest
to be skipped.


git-svn-id: https://svn.cern.ch/reps/psi46expert/trunk@42 4525493e-7705-40b1-a816-d608a930855b