Fixed the function in the 'Roc' class that reads the trim parameters from the textfile.
Previously it expected a strict ordering of the pixels in this file
and completely ignored the pixel address that is written there.

Now the function actually reads the pixel addresses and sets the trim bits
accordingly and if the pixel is not mentioned in the file it assumes
a value of 15. This behaviour is backward compatible.

The implementation is also more safe now.


git-svn-id: https://svn.cern.ch/reps/psi46expert/trunk@48 4525493e-7705-40b1-a816-d608a930855b