Checking for libusb-1.0 headers in the ./configure script and if they are not found the feature of disabling usb serial kernel modules for the testboards will be disabled.
This will prevent compilation errors that occur if although the
libusb-1.0 is installed the header files are missing because on
some systems they are separate from the binaries in a -devel
package.

The reason for not requiring these headers for compilation is
that on some systems one may not have sufficient access to
install such a system package.


git-svn-id: https://svn.cern.ch/reps/psi46expert/trunk@72 4525493e-7705-40b1-a816-d608a930855b