Corrected rate measurement for the high rate efficiency test. Edge double columns are now ignored and since it was found that the double column under test shows a much higher rate it is now ignored too.
The increased rate for the double column under test can be explained
qualitatively as follows:
Whenever a timestamp buffer overflow occurs hits from following or even
preceding drains can be assigned to the same timestamp. This can be called
hit migration. This occurs with x-ray as well as with calibrate hits. The
problem is that with calibrate hits it even occurs when there is no x-ray
hit and this leads to a higher amount of (migrated) x-ray hits to be
measured. The change in rate can be a factor greater than 2.


git-svn-id: https://svn.cern.ch/reps/psi46expert/branches/xray@97 4525493e-7705-40b1-a816-d608a930855b